restructure hash join modules
